Maintain: Optimize mean operation unit.

+++ include/filter/mean.h
@@ -7,15 +7,15 @@
 #include <mm/gfp.h>
 
 struct mean_state {
-    unsigned int count;
-    unsigned int mask;
-    long gross;
-    long buffer[0];
+    int count;
+    int mask;
+    int gross;
+    int buffer[0];
 };
 
-extern long mean_update(struct mean_state *state, long value);
+extern int mean_update(struct mean_state *state, int value);
 extern void mean_clear(struct mean_state *state);
-extern struct mean_state *mean_alloc(unsigned int length, gfp_t flags);
+extern struct mean_state *mean_alloc(int length, gfp_t flags);
 extern void mean_free(struct mean_state *state);
 
 #endif  /* _FILTER_MEAN_H_ */
