+++ usr/selftest/filter/mean.c
@@ -14,12 +14,12 @@ static state mean_testing(struct kshell_context *ctx, void *pdata)
 {
     struct mean_state *test = pdata;
     unsigned int count;
-    long rawdata, filtered;
+    short rawdata, filtered;
 
     for (count = 0; count < TEST_LOOP; ++count) {
-        rawdata = (long)prandom_value();
+        rawdata = (short)prandom_value();
         filtered = mean_update(test, rawdata);
-        kshell_printf(ctx, "mean filter test%02d: %11ld -> %11ld\n",
+        kshell_printf(ctx, "mean filter test%02d: %6d -> %6d\n",
                       count, rawdata, filtered);
     }
